requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-google-genai
chromadb
openai
streamlit
pypdf
tiktoken
python-dotenv
pandas
numpy
tqdm
